Header component with WalletConnnect
