+++ src/app/components/Header.tsx
@@ -1,3 +1,5 @@
+import WalletConnectButton from "./WalletConnectButton"
+
 export default function Header() {
     return (
         <>
@@ -9,7 +11,9 @@ export default function Header() {
                             <span className="self-center text-xl font-semibold whitespace-nowrap text-gray-700">Flowbite</span>
                         </a>
                         <div className="flex items-center lg:order-2">
-                            <a href="#" className="text-gray-800 hover:bg-gray-300 focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-4 lg:px-5 py-2 lg:py-2.5 mr-2"> Conectar Billetera </a>
+                            <a href="#" className="text-gray-800 hover:bg-gray-300 focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-4 lg:px-5 py-2 lg:py-2.5 mr-2"> 
+                                <WalletConnectButton /> 
+                            </a>
                             <a href="#" className="text-white bg-blue-500 hover:bg-gray-300 hover:text-gray-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-4 lg:px-5 py-2 lg:py-2.5 mr-2">Empezar</a>
                         </div>
                         <div className="hidden justify-between items-center w-full lg:flex lg:w-auto lg:order-1" id="mobile-menu-2">
